package @vexl-next/rest-api should not depend on @vexl-next/server-utils
