+++ apps/btc-exchange-rate-service/src/httpServer.ts
@@ -4,11 +4,11 @@ import {
   HttpMiddleware,
   HttpServer,
 } from '@effect/platform/index'
-import {ServerSecurityMiddlewareLive} from '@vexl-next/rest-api/src/apiSecurity'
 import {BtcExchangeRateApiSpecification} from '@vexl-next/rest-api/src/services/btcExchangeRate/specification'
 import {healthServerLayer} from '@vexl-next/server-utils/src/HealthServer'
 import {NodeHttpServerLiveWithPortFromEnv} from '@vexl-next/server-utils/src/NodeHttpServerLiveWithPortFromEnv'
 import {ServerCrypto} from '@vexl-next/server-utils/src/ServerCrypto'
+import {ServerSecurityMiddlewareLive} from '@vexl-next/server-utils/src/serverSecurity'
 import {Layer} from 'effect'
 import {cryptoConfig, healthServerPortConfig} from './configs'
 import {getExchangeRateHandler} from './handlers'
