+++ apps/contact-service/src/__tests__/routes/clubs/moderator/deactivateClubJoinLink.test.ts
@@ -8,13 +8,13 @@ import {NotFoundError} from '@vexl-next/domain/src/general/commonErrors'
 import {type ExpoNotificationToken} from '@vexl-next/domain/src/utility/ExpoNotificationToken.brand'
 import {UriStringE} from '@vexl-next/domain/src/utility/UriString.brand'
 import {
+  InvalidChallengeError,
+  type SignedChallenge,
+} from '@vexl-next/rest-api/src/challenges/contracts'
+import {
   InviteCodeNotFoundError,
   UserIsNotModeratorError,
 } from '@vexl-next/rest-api/src/services/contact/contracts'
-import {
-  InvalidChallengeError,
-  type SignedChallenge,
-} from '@vexl-next/server-utils/src/services/challenge/contracts'
 import {expectErrorResponse} from '@vexl-next/server-utils/src/tests/expectErrorResponse'
 import {addTestHeaders} from '@vexl-next/server-utils/src/tests/nodeTestingApp'
 import {Effect, Option, Schema} from 'effect'
