+++ apps/contact-service/src/__tests__/routes/clubs/moderator/listClubLinks.test.ts
@@ -4,11 +4,11 @@ import {generateClubUuid} from '@vexl-next/domain/src/general/clubs'
 import {NotFoundError} from '@vexl-next/domain/src/general/commonErrors'
 import {type ExpoNotificationToken} from '@vexl-next/domain/src/utility/ExpoNotificationToken.brand'
 import {UriStringE} from '@vexl-next/domain/src/utility/UriString.brand'
-import {UserIsNotModeratorError} from '@vexl-next/rest-api/src/services/contact/contracts'
 import {
   InvalidChallengeError,
   type SignedChallenge,
-} from '@vexl-next/server-utils/src/services/challenge/contracts'
+} from '@vexl-next/rest-api/src/challenges/contracts'
+import {UserIsNotModeratorError} from '@vexl-next/rest-api/src/services/contact/contracts'
 import {expectErrorResponse} from '@vexl-next/server-utils/src/tests/expectErrorResponse'
 import {addTestHeaders} from '@vexl-next/server-utils/src/tests/nodeTestingApp'
 import {Array, Effect, Option, Order, Schema} from 'effect'
