+++ apps/mobile/src/state/marketplace/domain.ts
@@ -22,7 +22,7 @@ import {type BasicError} from '@vexl-next/domain/src/utility/errors'
 import {type CryptoError} from '@vexl-next/generic-utils/src/effect-helpers/crypto'
 import {type ApiErrorFetchingClubsOffers} from '@vexl-next/resources-utils/src/offers/getNewClubsOffersAndDecrypt'
 import {type ApiErrorFetchingOffers} from '@vexl-next/resources-utils/src/offers/getNewOffersAndDecrypt'
-import {type ErrorSigningChallenge} from '@vexl-next/server-utils/src/services/challenge/contracts'
+import {type ErrorSigningChallenge} from '@vexl-next/rest-api/src/challenges/contracts'
 import {Schema} from 'effect'
 import {z} from 'zod'
 import {fastDeepEqualRemoveUndefineds} from '../../utils/fastDeepEqualRemoveUndefineds'
