+++ apps/user-service/src/__tests__/login/loginFlow.test.ts
@@ -2,8 +2,8 @@ import {generatePrivateKey} from '@vexl-next/cryptography/src/KeyHolder'
 import {ecdsaSign} from '@vexl-next/cryptography/src/operations/ecdsa'
 import {E164PhoneNumberE} from '@vexl-next/domain/src/general/E164PhoneNumber.brand'
 import {EcdsaSignature} from '@vexl-next/generic-utils/src/effect-helpers/crypto'
-import {verifyUserSecurity} from '@vexl-next/rest-api/src/apiSecurity'
 import {CommonHeaders} from '@vexl-next/rest-api/src/commonHeaders'
+import {verifyUserSecurity} from '@vexl-next/server-utils/src/serverSecurity'
 import {mockedReportNewUserCreated} from '@vexl-next/server-utils/src/tests/mockedDashboardReportsService'
 import {Effect, pipe, Schema} from 'effect'
 import {LoggedInUsersDbService} from '../../db/loggedInUsersDb'
