+++ packages/rest-api/package.json
@@ -28,7 +28,6 @@
     "@effect/platform": "^0.92.1",
     "@vexl-next/cryptography": "0.0.0",
     "@vexl-next/domain": "0.0.0",
-    "@vexl-next/server-utils": "0.0.0",
     "effect": "^3.18.4",
     "fast-check": "^3.20.0",
     "fp-ts": "^2.13.1",
