+++ packages/rest-api/src/challenges/contracts.ts
@@ -1,18 +1,6 @@
 import {PublicKeyPemBase64E} from '@vexl-next/cryptography/src/KeyHolder'
 import {EcdsaSignature} from '@vexl-next/generic-utils/src/effect-helpers/crypto'
-import {Schema} from 'effect'
-
-export class InvalidChallengeError extends Schema.TaggedError<InvalidChallengeError>(
-  'InvalidChallengeError'
-)('InvalidChallengeError', {
-  status: Schema.optionalWith(Schema.Literal(400), {default: () => 400}),
-}) {}
-
-export class ErrorSigningChallenge extends Schema.TaggedError<ErrorSigningChallenge>(
-  'ErrorSigningChallenge'
-)('ErrorSigningChallenge', {
-  status: Schema.optionalWith(Schema.Literal(400), {default: () => 400}),
-}) {}
+import {Schema} from 'effect/index'
 
 export const Challenge = Schema.String.pipe(Schema.brand('Challenge'))
 export type Challenge = Schema.Schema.Type<typeof Challenge>
@@ -28,3 +16,15 @@ export const RequestBaseWithChallenge = Schema.Struct({
   signedChallenge: SignedChallenge,
 })
 export type RequestBaseWithChallenge = typeof RequestBaseWithChallenge.Type
+
+export class InvalidChallengeError extends Schema.TaggedError<InvalidChallengeError>(
+  'InvalidChallengeError'
+)('InvalidChallengeError', {
+  status: Schema.optionalWith(Schema.Literal(400), {default: () => 400}),
+}) {}
+
+export class ErrorSigningChallenge extends Schema.TaggedError<ErrorSigningChallenge>(
+  'ErrorSigningChallenge'
+)('ErrorSigningChallenge', {
+  status: Schema.optionalWith(Schema.Literal(400), {default: () => 400}),
+}) {}
