+++ packages/rest-api/src/challenges/specification.ts
