+++ packages/server-utils/package.json
@@ -38,6 +38,7 @@
     "@vexl-next/cryptography": "0.0.0",
     "@vexl-next/domain": "0.0.0",
     "@vexl-next/generic-utils": "0.0.0",
+    "@vexl-next/rest-api": "0.0.0",
     "awesome-phonenumber": "^7.4.0",
     "bullmq": "^5.52.1",
     "effect": "^3.18.4",
